Minor improvements and fixes, including a change that makes the moralis resync a little more resilient to api rate limit problems.

+++ helperclass_timer.py
@@ -10,7 +10,7 @@ class Timer:
         limits, but I suspect it will grow to have other functionality as well.
     """
 
-    def __init__(self, rate_limit: int = 5):
+    def __init__(self, rate_limit: float = 5):
         """
         Initialize an instance of the class.
           Args:
